+++ CHANGELOG.md
@@ -3,6 +3,14 @@
   Placeholder for the next version (at the beginning of the line):
   ## **WORK IN PROGRESS**
 -->
+## **WORK IN PROGRESS**
+
+- Import/Export of Screens/Images/Controls
+- Define your own Controls directly in webui
+- Drag/Drop of Icons/Images to Properties
+- Drag/Drop of objects to Bindings-Editor Signalname
+- Basic functionality of CustomControls
+
 ## 0.1.0 (2023-08-27)
 
 -   initial public release